# swagger.yaml
openapi: 3.0.3
info:
  title: Helpp API
  description: |
    API do projeto **Helpp** — plataforma de ajuda mútua que conecta pessoas que precisam de ajuda com quem pode ajudar.

    ## Autenticação
    A maioria dos endpoints protegidos usa **Bearer Token (JWT)**. O access token expira em **15 minutos**.
    Use `POST /auth/refresh` para renová-lo com o refresh token (validade de 30 dias).

    ## Sistema de Reciprocidade
    Cada usuário possui um saldo de reciprocidade:
    - `helpReceived` incrementa quando recebe uma ajuda concluída
    - `helpGiven` incrementa quando conclui uma ajuda como helper
    - Se `helpReceived > helpGiven`, o usuário está bloqueado de criar novos pedidos até ajudar alguém
  version: 1.0.0
  contact:
    name: Helpp
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        nickname:
          type: string
        cpfSuffix:
          type: string
          description: Últimos 3 dígitos do CPF
        phoneSuffix:
          type: string
          description: Últimos 4 dígitos do telefone
        city:
          type: string
        neighborhood:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        avatar:
          type: string
          description: Imagem em base64
        profileComplete:
          type: boolean
        helpReceived:
          type: integer
          description: Total de ajudas recebidas
          default: 0
        helpGiven:
          type: integer
          description: Total de ajudas dadas
          default: 0
    Request:
      type: object
      properties:
        _id:
          type: string
        userId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        category:
          type: string
          enum:
            - Food or Groceries
            - Medicine
            - Transportation
            - Temporary Shelter
            - Help with Bills
            - Job Opportunity
            - School Supplies
            - Emotional Support
            - Other
        reason:
          type: string
        urgency:
          type: string
          enum: [Low, Medium, High, Emergency]
        giveBack:
          type: string
        status:
          type: string
          enum: [pending, matched, completed, cancelled]
          default: pending
        latitude:
          type: number
        longitude:
          type: number
        createdAt:
          type: string
          format: date-time
...
    Sponsor:
      type: object
      properties:
        _id:
          type: string
        companyName:
          type: string
        cnpj:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        supportType:
          type: string
        notes:
          type: string
        logo:
          type: string
          description: Imagem em base64
    SponsorRequest:
      type: object
      properties:
        _id:
          type: string
        sponsorId:
          type: string
        productId:
          type: string
        productName:
          type: string
        companyName:
          type: string
        userId:
          type: string
        userName:
          type: string
        userContact:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, contacted, resolved]
          default: pending
        contactedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    ObligationPendingError:
      type: object
      properties:
        code:
          type: string
          example: obligation_pending
        helpReceived:
          type: integer
          example: 2
        helpGiven:
          type: integer
          example: 1
paths:
...
  /requests:
    get:
      tags: [Requests]
      summary: Lista pedidos abertos
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'

    post:
      tags: [Requests]
      summary: Cria novo pedido de ajuda
      description: |
        Cria um novo pedido. **Bloqueado** se o usuário tiver obrigação pendente (`helpReceived > helpGiven`).
        Nesse caso retorna `403` com `code: "obligation_pending"`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, reason, urgency, giveBack]
              properties:
                category:
                  type: string
                  enum:
                    - Food or Groceries
                    - Medicine
                    - Transportation
                    - Temporary Shelter
                    - Help with Bills
                    - Job Opportunity
                    - School Supplies
                    - Emotional Support
                    - Other
                reason:
                  type: string
                urgency:
                  type: string
                  enum: [Low, Medium, High, Emergency]
                giveBack:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Campos obrigatórios faltando
        '403':
          description: Obrigação de reciprocidade pendente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObligationPendingError'
  /requests/mine:
    get:
      tags: [Requests]
      summary: Lista pedidos do usuário autenticado (com matches e offers)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pedidos do usuário com informações de match
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Request'
                    - type: object
                      properties:
                        matchId:
                          type: string
                        matchStatus:
                          type: string
                        helper:
                          type: object
                        offers:
                          type: array
                          items:
                            type: object
  /requests/{id}:
    get:
      tags: [Requests]
      summary: Detalhe de um pedido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '404':
          description: Pedido não encontrado

    patch:
      tags: [Requests]
      summary: Atualiza status do pedido
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, matched, completed, cancelled]
      responses:
        '200':
          description: Status atualizado
        '403':
          description: Acesso negado
        '404':
          description: Pedido não encontrado
...
  /sponsors:
    get:
      tags: [Sponsors]
      summary: Lista todos os apoiadores
      responses:
        '200':
          description: Lista de apoiadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sponsor'

    post:
      tags: [Sponsors]
      summary: Cadastra novo apoiador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyName, cnpj, contactName, email, city]
              properties:
                companyName:
                  type: string
                cnpj:
                  type: string
                contactName:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                city:
                  type: string
                supportType:
                  type: string
                notes:
                  type: string
                logo:
                  type: string
                  description: Imagem em base64
      responses:
        '201':
          description: Apoiador cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '409':
          description: CNPJ já cadastrado
  /sponsors/{id}:
    get:
      tags: [Sponsors]
      summary: Busca apoiador por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Apoiador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '404':
          description: Apoiador não encontrado
  /sponsors/{id}/logo:
    patch:
      tags: [Sponsors]
      summary: Atualiza logo do apoiador
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logo:
                  type: string
                  nullable: true
                  description: Base64 ou null para remover
      responses:
        '200':
          description: Logo atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  logo:
                    type: string
                    nullable: true
  /sponsor-requests:
    get:
      tags: [Sponsor Requests]
      summary: Lista requisições de apoio (filtros por sponsorId ou userId)
      parameters:
        - in: query
          name: sponsorId
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
      responses:
        '200':
          description: Lista de requisições
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SponsorRequest'

    post:
      tags: [Sponsor Requests]
      summary: Cria requisição de apoio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sponsorId, productId, productName, userName, userContact]
              properties:
                sponsorId:
                  type: string
                productId:
                  type: string
                productName:
                  type: string
                companyName:
                  type: string
                userId:
                  type: string
                userName:
                  type: string
                userContact:
                  type: string
                message:
                  type: string
      responses:
        '201':
          description: Requisição criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorRequest'
  /sponsor-requests/{id}/status:
    patch:
      tags: [Sponsor Requests]
      summary: Atualiza status da requisição
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, contacted, resolved]
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorRequest'
        '404':
          description: Requisição não encontrada
...